website: PMTiles server and f16 shader support
The PMTiles server is a Cloudflare worker, not included in this repo,
but it allows fetching the .pmtile data with `z/x/y` URLs. These are
normal HTTP requests (as opposed to HTTP range requests) and so can be
cached by browsers, CDNs etc.

The `f16` shader support is a fix to allow browsers that don't support
single `f32` shader textures to still process `f32` data via packed data
in the red and green channels of a shader texture. This should fix a few
browsers that couldn't display the heatmap, most notably Chrome.